Renamed package that has mappers that deal with JSON files to more appropriately match what they are for.  Started Object Relational Mappers that will serve as the database model.
