+++ Model/JsonMappers/ConferenceMapper.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace HalfboardStats.Model.ObjectRelationalMappers
+namespace HalfboardStats.Model.JsonMappers
 {
-    public class DivisionMapper
+    public class ConferenceMapper
     {
         public int Id { get; set; }
         public string Name { get; set; }
