+++ Model/JsonMappers/DivisionMapper.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace HalfboardStats.Model.JsonMappers
+{
+    public class DivisionMapper
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public string Link { get; set; }
+    }
+}
