+++ Model/JsonMappers/LeagueRecordMapper.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace HalfboardStats.Model.ObjectRelationalMappers
+namespace HalfboardStats.Model.JsonMappers
 {
     public class LeagueRecordMapper
     {
