+++ Model/JsonMappers/TeamMapper.cs
@@ -3,12 +3,13 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace HalfboardStats.Model.ObjectRelationalMappers
+namespace HalfboardStats.Model.JsonMappers
 {
-    public class LeagueMapper
+    public class TeamMapper
     {
         public int Id { get; set; }
         public string Name { get; set; }
         public string Link { get; set; }
+
     }
 }
